+++ .gitignore
@@ -1,3 +1,3 @@
 .pytest_cache
 report_allure
-creds.py
+support/creds.py
