+++ support/logging.py
@@ -1,0 +1,5 @@
+from loguru import logger
+
+class Logger():
+    logger.remove()
+    logger.add("debug.log", format="{time:DD:MM:YYYY:h:m:s} {level} {message}", level="DEBUG")
